Add a reasonable default Redis persistence
Despite the claim of the README.md, `PostgreSQL and Redis data is persisted in Docker volumes`, this is not actually the case. The volume is present, but Redis is not configured to use it.

This question comes up on the forum monthly, if not more frequently. Choose a reasonable default for people to build on.